feat(scripts): Change logic
The script will not show "Bypass is working" toast anymore. It will only show a toast if the bypass isn't working or if the bypass status can't be determined which ensures that you don't get annoyed by the repetitive "Bypass is working" dialog while keeping the core script functionality working. So, no need to worry if you don't see any "Bypass is working" toast after enabling the script(as in the video guide), rest assured, the bypass is working.

Thanks to @Brokoli5191